The full split of connection tests
